--- a/src/daft-json/src/schema.rs
+++ b/src/daft-json/src/schema.rs
@@ -166,7 +166,7 @@ where
             let mut record = record.context(StdIOSnafu)?;
 
             // Parse record into a JSON Value, then infer the schema.
-            let parsed_record = simd_json::to_borrowed_value(unsafe { record.as_bytes_mut() })
+            let parsed_record = crate::deserializer::to_value(unsafe { record.as_bytes_mut() })
                 .map_err(|e| super::Error::JsonDeserializationError {
                     string: e.to_string(),
                 })?;